Split the backend contract tests into their own file
The structure gate refuses a source file past five hundred lines, and
the registry tests carry enough scenario text to put backend.rs over it.
The tests move to backend/tests.rs unchanged; the module path they
compile under is the same, so every `super::` and cross-module path
still resolves.
